Refactor Level index
